chore(event-bus): event bus error handling
