+++ .changeset/orange-donkeys-hammer.md
@@ -1,0 +1,8 @@
+---
+"@medusajs/event-bus-local": patch
+"@medusajs/event-bus-redis": patch
+"@medusajs/orchestration": patch
+"@medusajs/utils": patch
+---
+
+Improve event bus error handling
