+++ packages/core/orchestration/src/transaction/errors.ts
@@ -58,30 +58,3 @@ export class TransactionTimeoutError extends Error {
     this.name = "TransactionTimeoutError"
   }
 }
-
-export function serializeError(error) {
-  const serialized = {
-    message: error.message,
-    name: error.name,
-    stack: error.stack,
-  }
-
-  Object.getOwnPropertyNames(error).forEach((key) => {
-    // eslint-disable-next-line no-prototype-builtins
-    if (!serialized.hasOwnProperty(key)) {
-      serialized[key] = error[key]
-    }
-  })
-
-  return serialized
-}
-
-export function isErrorLike(value) {
-  return (
-    !!value &&
-    typeof value === "object" &&
-    "name" in value &&
-    "message" in value &&
-    "stack" in value
-  )
-}
