+++ packages/core/utils/src/common/index.ts
@@ -18,6 +18,7 @@ export * from "./dynamic-import"
 export * from "./env-editor"
 export * from "./errors"
 export * from "./file-system"
+export * from "./filter-operator-map"
 export * from "./generate-entity-id"
 export * from "./get-caller-file-path"
 export * from "./get-config-file"
@@ -34,6 +35,7 @@ export * from "./is-boolean"
 export * from "./is-date"
 export * from "./is-defined"
 export * from "./is-email"
+export * from "./is-error-like"
 export * from "./is-object"
 export * from "./is-present"
 export * from "./is-string"
@@ -62,6 +64,7 @@ export * from "./remove-undefined-properties"
 export * from "./resolve-exports"
 export * from "./rules"
 export * from "./selector-constraints-to-string"
+export * from "./serialize-error"
 export * from "./set-metadata"
 export * from "./simple-hash"
 export * from "./string-to-select-relation-object"
@@ -74,4 +77,3 @@ export * from "./trim-zeros"
 export * from "./upper-case-first"
 export * from "./validate-handle"
 export * from "./wrap-handler"
-export * from "./filter-operator-map"
