+++ packages/core/utils/src/common/is-error-like.ts
@@ -1,0 +1,9 @@
+export function isErrorLike(value) {
+  return (
+    !!value &&
+    typeof value === "object" &&
+    "name" in value &&
+    "message" in value &&
+    "stack" in value
+  )
+}
